refactor: move contact mapper to screen state file and rename `onInit` function
